pyproject.toml
[tool.poetry]
name = "toogoodtogoreminder"
version = "0.1.0"
description = ""
authors = ["Karol Puchała <puchala1.karol@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tgtg = "^0.16.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
